+++ src/bufkit_data/surface_section.rs
@@ -247,49 +247,49 @@ mod test {
             surface_section
                 .into_iter()
                 .map(|sd| sd.mslp)
-                .collect::<Vec<f32>>(),
+                .collect::<Vec<f64>>(),
             vec![1020.4, 1021.5, 1022.3, 1022.7, 1022.6, 1021.8]
         );
         assert_eq!(
             surface_section
                 .into_iter()
                 .map(|sd| sd.station_pres)
-                .collect::<Vec<f32>>(),
+                .collect::<Vec<f64>>(),
             vec![909.1, 909.4, 909.4, 909.3, 908.8, 908.6]
         );
         assert_eq!(
             surface_section
                 .into_iter()
                 .map(|sd| sd.low_cloud)
-                .collect::<Vec<f32>>(),
+                .collect::<Vec<f64>>(),
             vec![100.0, 52.0, 2.0, 1.0, 3.0, 4.0]
         );
         assert_eq!(
             surface_section
                 .into_iter()
                 .map(|sd| sd.mid_cloud)
-                .collect::<Vec<f32>>(),
+                .collect::<Vec<f64>>(),
             vec![0.0; 6]
         );
         assert_eq!(
             surface_section
                 .into_iter()
                 .map(|sd| sd.hi_cloud)
-                .collect::<Vec<f32>>(),
+                .collect::<Vec<f64>>(),
             vec![58.0, 61.0, 39.0, 33.0, 49.0, 77.0]
         );
         assert_eq!(
             surface_section
                 .into_iter()
                 .map(|sd| sd.uwind)
-                .collect::<Vec<f32>>(),
+                .collect::<Vec<f64>>(),
             vec![0.9, 1.3, 1.1, 1.1, 0.6, 0.4]
         );
         assert_eq!(
             surface_section
                 .into_iter()
                 .map(|sd| sd.vwind)
-                .collect::<Vec<f32>>(),
+                .collect::<Vec<f64>>(),
             vec![-0.1, 0.5, 0.6, 0.6, 0.8, 0.6]
         );
     }
